// src/app/interfaces/Employee.ts
export interface Employee {
    id: string;
    name: string;
    cpf: string;
    birthDate: string;
    email: string;
    phoneNumber: string;
    homeNumber: string;
    logradouro: string;
    complementResidential: string;
    neighborhood: string;
    city: string;
    state: string;
    type: string;
    status: string;
    gender: string;
}